requirements.txt
# Azure OpenAI
openai>=1.0.0
azure-identity>=1.15.0

# Environment management
python-dotenv>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Logging and utilities
rich>=13.0.0
colorlog>=6.7.0

# Configuration
pydantic>=2.0.0
pyyaml>=6.0

# Optional: for visualization
matplotlib>=3.7.0
seaborn>=0.12.0
